Create aws_tg_instances_terminate_v1.sh
In this updated version, instead of using tags, we modify the instance description using the aws ec2 modify-instance-attribute command. We pass the captured instance name ($INSTANCE_NAME) as the value of the description attribute for the new instance ($NEW_INSTANCE_ID). This will set the instance description to the captured instance name.